refactor: clean up main and create proper modules for peripherals
